backend/requirements.txt
fastapi
uvicorn
pymilvus
pymongo
langchain
langchain-community
langchain-experimental
langchain-text-splitters
langchain-huggingface
sentence-transformers
torch
nltk
python-dotenv
pytesseract
pillow
pdf2image
PyPDF2
langchain 
langchain-community 
langchain-text-splitters 
pymilvus 
pypdf 
python-docx 
unstructured 
unstructured[pdf] 
pillow 
pytesseract 
opencv-python 
sentence-transformers 
python-multipart